requirements.txt
pynput
pycryptodome
